+++ src/cassandra_statement.rs
@@ -385,8 +385,6 @@ impl Display for CassandraStatement {
 #[cfg(test)]
 mod tests {
     use crate::cassandra_ast::CassandraAST;
-    use crate::cassandra_statement::CassandraStatement;
-    use crate::select::{Named, SelectElement};
 
     // only tests single results
     fn test_parsing(expected: &[&str], statements: &[&str]) {
@@ -1294,103 +1292,5 @@ mod tests {
             "ALTER MATERIALIZED VIEW keyspace.mview WITH option1 = 'option' AND option2 = 3.5",
         ];
         test_parsing(&expected, &stmts);
-    }
-
-    #[test]
-    fn test_invalid_statement() {
-        let statement = "This is an invalid statement";
-        let ast = CassandraAST::new(statement);
-        assert!(ast.has_error());
-        let result = &ast.statements[0];
-        matches!(result.statement, CassandraStatement::Unknown(_));
-        assert!(result.has_error);
-        assert_eq!(statement.to_string(), result.statement.to_string());
-        assert_eq!(
-            statement,
-            match &result.statement {
-                CassandraStatement::Unknown(text) => text,
-                _ => "",
-            }
-        );
-    }
-
-    #[test]
-    fn test_partial_invalid_statement() {
-        let statement = "SELECT * FROM foo WHERE some invalid part";
-        let expected = ["SELECT * FROM foo", "WHERE some invalid part"];
-        let ast = CassandraAST::new(statement);
-        assert!(ast.has_error());
-
-        let result = &ast.statements[0];
-        assert!(!result.has_error);
-        matches!(result.statement, CassandraStatement::Select(_));
-        assert_eq!(expected[0].to_string(), result.statement.to_string());
-
-        let result = &ast.statements[1];
-        assert!(result.has_error);
-        matches!(result.statement, CassandraStatement::Unknown(_));
-        assert_eq!(expected[1].to_string(), result.statement.to_string());
-        assert_eq!(
-            expected[1],
-            match &result.statement {
-                CassandraStatement::Unknown(text) => text,
-                _ => "",
-            }
-        );
-    }
-
-    #[test]
-    fn test_multiple_statements() {
-        let stmt = "Select * from foo; Select * from bar;";
-        let ast = CassandraAST::new(stmt);
-        assert!(!ast.has_error());
-        assert_eq!(2, ast.statements.len());
-    }
-
-    #[test]
-    fn test_unicode_chars() {
-        let stmt = "SELECT * FROM foo WHERE bar = '\u{1F44D}'";
-        let ast = CassandraAST::new(stmt);
-        assert!(!ast.has_error());
-        let result = &ast.statements[0];
-        assert!(!result.has_error);
-        assert_eq!(stmt.to_string(), result.statement.to_string());
-    }
-
-    #[test]
-    fn test_select_element_display() {
-        assert_eq!("*", SelectElement::Star.to_string());
-        assert_eq!(
-            "col",
-            SelectElement::Column(Named {
-                name: "col".to_string(),
-                alias: None
-            })
-            .to_string()
-        );
-        assert_eq!(
-            "func",
-            SelectElement::Function(Named {
-                name: "func".to_string(),
-                alias: None
-            })
-            .to_string()
-        );
-        assert_eq!(
-            "col AS alias",
-            SelectElement::Column(Named {
-                name: "col".to_string(),
-                alias: Some("alias".to_string())
-            })
-            .to_string()
-        );
-        assert_eq!(
-            "func AS alias",
-            SelectElement::Function(Named {
-                name: "func".to_string(),
-                alias: Some("alias".to_string())
-            })
-            .to_string()
-        );
     }
 }
